Use distribution license as dataset license
When importing a dataset and related distributions (aka resources in
CKAN) from an RDF DCAT source, license information is on distributions.
However in CKAN, this information is on dataset.

So we pick one distribution and use its license as a license for the
dataset. This is only done if a matching license if found in CKAN
licence register.

This is a weak method since distributions may have differing
licenses, even for one dataset.